fix: fixing wrong types in the tmdb-search-movies

+++ src/graphql/type-defs/schemas/tmdb/search-movies.schema.ts
@@ -7,12 +7,12 @@ export default `#graphql
     originalLanguage: String
     originalTitle: String
     overview: String
-    popularity: Int
+    popularity: Float
     posterPath: String
     releaseDate: String
     title: String
     video: Boolean
-    voteAverage: Int
+    voteAverage: Float
     voteCount: Int
   }
